Verification of Create Simple Todo request body
