+++ src/backend/service/simpleTodoService.go
@@ -1,38 +1,43 @@
 package service
 
 import (
+	"encoding/json"
 	"fmt"
-	"main/constants"
 	"main/util"
 	"net/http"
+	"regexp"
 
-	"github.com/dgrijalva/jwt-go"
 	"github.com/julienschmidt/httprouter"
 )
 
+type SimpleTodo struct {
+	Description string `json:"description"`
+	TargetDate  string `json:"targetDate"`
+}
+
+// @TODO: Need to get the username from decoding JWT after verifying it
+
 func HandleCreateSimpleTodo(writer http.ResponseWriter, req *http.Request, _ httprouter.Params) {
-	if err := verifyJWT(req.Header.Get("JWT")); err != nil {
+	if err := util.VerifyJWT(req.Header.Get("JWT")); err != nil {
 		util.ThrowUnauthorisedRequest(writer)
 		return
 	}
 
+	var todo SimpleTodo
+	json.NewDecoder(req.Body).Decode(&todo)
+
+	if err := verifyCreateTodoRequestBody(todo); err != nil {
+		util.ThrowBadRequest(writer, err.Error())
+	}
+
 }
 
-func verifyJWT(tokenString string) error {
-	token, err := jwt.Parse(tokenString, func(token *jwt.Token) (interface{}, error) {
-		return SECRET_KEY, nil
-	})
-	if err != nil {
-		return err
+func verifyCreateTodoRequestBody(todo SimpleTodo) error {
+	if len(todo.Description) == 0 {
+		return fmt.Errorf("description field must not be empty")
 	}
-	if !token.Valid {
-		return fmt.Errorf(constants.INVALID_JWT)
-	}
-
-	_, ok := token.Claims.(jwt.MapClaims)
-
-	if !ok {
-		return fmt.Errorf(constants.INVALID_CLAIMS)
+	if match, _ := regexp.MatchString("^\\d{4}-\\d{2}-\\d{2}$", todo.TargetDate); !match {
+		return fmt.Errorf("targetDate field must be of the format YYYY-MM-DD")
 	}
 
 	return nil
